% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConf}
\alias{defConf}
\title{defConf}
\usage{
defConf(
  matern_intensity = 1,
  mmpp = 1,
  detectionTrunc = -1,
  matern_size = 1,
  g_function = 1,
  covRate = "",
  covDetection = "",
  usePCpriors = 0,
  pcPriorsRange_intensity = c(300, 0.1),
  pcPriorsSD_intensity = c(1, 0.1),
  pcPriorsRange_size = c(300, 0.1),
  pcPriorsSD_size = c(1, 0.1),
  penalizeMMPP = c(1, 0.1),
  spdeDetails = list(cutoff = 50, max.edge = c(80, 150)),
  applyPodSize = 1,
  podSizeDist = 2,
  independentPodSize = 1,
  dependentPodSize = c(0, 1),
  cellsize = 20,
  UTMproj = CRS("+proj=utm +zone=22 +units=km"),
  LatLonProj = CRS("+proj=longlat +datum=WGS84")
)
}
\arguments{
\item{matern_intensity}{1: include spatial effect in group abundance intensity, 0: not}

\item{mmpp}{1: include MMPP in group abundance intensity, 0:not}

\item{detectionTrunc}{Distance at which detection is truncated}

\item{matern_size}{1: include spatial effect in group size, 0: not}

\item{g_function}{1: Half normal detection function, 2: Hazard detection function}

\item{covRate}{Covariates in group abundance intensity}

\item{covDetection}{Covariates in detection function}

\item{usePCpriors}{Use pc-priors (penalty) for spatial effect model parameters}

\item{pcPriorsRange_intensity}{pc-prior group abundance intensity range}

\item{pcPriorsSD_intensity}{pc-prior group abundance intensity marginal standard deviation}

\item{pcPriorsRange_size}{pc-prior group size intensity range}

\item{pcPriorsSD_size}{pc-prior group size intensity marginal standard deviation}

\item{spdeDetails}{Details for constructing mesh with use of INLA}

\item{applyPodSize}{1: Pod size is included, 0: Not included}

\item{podSizeDist}{1:poisson, 2:negative binomial group size}

\item{independentPodSize}{1: Size independent of intensity}

\item{dependentPodSize}{First if beta0 is included and second if beta_SPDE is included}

\item{cellsize}{Distance between spatial quadratic integration points for total abundance}

\item{UTMproj}{CRS-UTM projection}

\item{LatLonProj}{CRS-latitude and longitude projection}
}
\value{
Configurations
}
\description{
defConf
}
