% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_functions.R
\name{eshw_half_normal}
\alias{eshw_half_normal}
\title{Half normal effective half width}
\usage{
eshw_half_normal(z, beta, w)
}
\arguments{
\item{z}{row from (covariate) design matrix}

\item{beta}{parameter vector (all parameters needed for g)}

\item{w}{truncation distance}
}
\value{
Half normal effective width
}
\description{
Half normal effective half width
}
