% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_functions.R
\name{eshw_hazard}
\alias{eshw_hazard}
\title{Hazard effective half width}
\usage{
eshw_hazard(z, beta, b, w)
}
\arguments{
\item{z}{row from (covariate) design matrix}

\item{beta}{parameter vector (all parameters needed for g)}

\item{b}{parameter in hazard detection function}

\item{w}{truncation distance}
}
\value{
Hazard effective width
}
\description{
Hazard effective half width
}
