% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{expm_mmpp2}
\alias{expm_mmpp2}
\title{expm_mmpp2 function}
\usage{
expm_mmpp2(lambda, mu, l)
}
\arguments{
\item{mu}{parameters in MMPP}

\item{l}{distance moved allong transect}

\item{lamdba}{Vector with low and high intensity}
}
\value{
Matrix exponential for MMPP with two states
}
\description{
Direct implementation for the matrix exponential in dim 2 for MMPP
}
