% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitLGPP}
\alias{fitLGPP}
\title{Fit the LGPP model}
\usage{
fitLGPP(data, par, conf, rel.tol = 1e-10, map = setMap(conf, par), ...)
}
\arguments{
\item{data}{data needed}

\item{par}{data needed}

\item{conf}{data needed}

\item{rel.tol}{data needed}

\item{map}{data needed}

\item{...}{sent to sdreport}
}
\value{
fitted LGPP model
}
\description{
Fit the LGPP model
}
