% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_functions.R
\name{g_half_normal}
\alias{g_half_normal}
\title{Half normal detection function}
\usage{
g_half_normal(y, z, beta)
}
\arguments{
\item{y}{perpendicular distance}

\item{z}{row from (covariate) design matrix}

\item{beta}{parameter vector (all parameters needed for g)}
}
\value{
Half normal detection density
}
\description{
Half normal detection function
}
