% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_functions.R
\name{g_hazard}
\alias{g_hazard}
\title{Hazard detection function}
\usage{
g_hazard(y, z, beta, b)
}
\arguments{
\item{y}{perpendicular distance}

\item{z}{row from (covariate) design matrix}

\item{beta}{parameter vector (all parameters needed for g)}

\item{b}{parameter in hazard detection}
}
\value{
Hazard detection density
}
\description{
Hazard detection function
}
