% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\title{jit}
\usage{
jit(run, sd = 0.2, nojit = 50, ncores = 2)
}
\arguments{
\item{run}{fitted model returned by \code{\link{fitMSLT}}}

\item{sd}{standard deviation in jitter analysis}

\item{ncores}{number of cores used in jitter analysis}
}
\value{
fitted model
}
\description{
jit
}
