% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.mslt}
\alias{logLik.mslt}
\title{logLik.mslt}
\usage{
\method{logLik}{mslt}(object, ...)
}
\arguments{
\item{object}{fitted model retuned by \code{\link{fitMSLT}}}

\item{...}{Extra argument}
}
\description{
logLik.mslt
}
\details{
Returns the log-likelihood and the number of parameters
}
