% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{partable}
\alias{partable}
\title{Return table with model parameters}
\usage{
partable(run, include_biasCorr_abundance = FALSE, include_range = FALSE)
}
\arguments{
\item{run}{run to extract model parameters from}
}
\value{
Table with model parameters
}
\description{
...
}
