% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{partable_model_derived}
\alias{partable_model_derived}
\title{Return table with model parameters}
\usage{
partable_model_derived(run, sdreport_run, CI_level = 0.95)
}
\arguments{
\item{run}{run to extract model parameters from}
}
\value{
Table with model parameters
}
\description{
...
}
