% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.jit}
\alias{print.jit}
\title{print.jit}
\usage{
\method{print}{jit}(jit, ...)
}
\arguments{
\item{jit}{List with the jitter runs}

\item{...}{Extra argument}
}
\description{
print.jit
}
\details{
Print results of the jitter analysis; Difference in parameters, abundance and log likelihood
}
