% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.lineTransectLGPP}
\alias{print.lineTransectLGPP}
\title{Print lineTransectLGPP object}
\usage{
\method{print}{lineTransectLGPP}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Print lineTransectLGPP object
}
\details{
Print log-likelihood and the main convergence criteria
}
