% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{Q_spde}
\alias{Q_spde}
\title{Q_spde function}
\usage{
Q_spde(spdeMatricesS, kappa)
}
\arguments{
\item{spdeMatricesS}{SPDE-matrices given by INLA needed to set up precision matrix}

\item{kappa}{spatial scale parameter}
}
\value{
Precision matrix for SPDE-procedure
}
\description{
Set up precision matrix for SPDE-procedure
}
