% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{readData}
\alias{readData}
\title{readData This provides the data in the format that the user is supposed to have}
\usage{
readData(maxDist = 1, minDist = 0)
}
\arguments{
\item{maxDist}{maximum distance between integration points}

\item{minDist}{minimum distance between integration points}
}
\value{
Processed data
}
\description{
...
}
