% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setData}
\alias{setData}
\title{setData}
\usage{
setData(d, predAreaUTM, conf)
}
\arguments{
\item{d}{Raw data}

\item{predAreaUTM}{prediciton area}

\item{conf}{configuration applied}
}
\value{
data used internally in package
}
\description{
Define data used internally based on data porvided and configuration file
}
