% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{setMap}
\alias{setMap}
\title{setMap}
\usage{
setMap(conf, par)
}
\arguments{
\item{conf}{Model configurations returned by \code{\link{defConf}}}

\item{par}{Model parameters set by \code{\link{setPar}}}
}
\value{
map-variable used by TMB
}
\description{
setMap
}
