% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{setPar}
\alias{setPar}
\title{setPar}
\usage{
setPar(data, conf)
}
\arguments{
\item{data}{data list returned by \code{\link{setData}}}

\item{conf}{configuration list returned by \code{\link{defConf}}}
}
\value{
Initial model parameters
}
\description{
setPar
}
