% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\title{jit}
\usage{
jit(run, sd = 0.2, nojit = 50, ncores = 2)
}
\arguments{
\item{run}{Fitted model returned by \code{\link{fitMSLT}}}

\item{sd}{Standard deviation in jitter analysis}

\item{nojit}{Number of sets of starting values}

\item{ncores}{Number of cores used in jitter analysis}
}
\value{
Fitted model
}
\description{
jit
}
