% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{partable}
\alias{partable}
\title{Return table with model parameters}
\usage{
partable(run, CI_level = 0.95)
}
\arguments{
\item{run}{run to extract model parameters from}

\item{CI_level}{Confidence level to accompany the point estimates with}
}
\value{
Table with model parameters
}
\description{
Return table with model parameters
}
