% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{partable_model_derived}
\alias{partable_model_derived}
\title{Return table with derived model parameters}
\usage{
partable_model_derived(run, sdreport_run = NULL, CI_level = 0.95)
}
\arguments{
\item{run}{run to extract model parameters from}

\item{sdreport_run}{the sdreport object for the run to extract model parameters from}

\item{CI_level}{Confidence level to accompany the point estimates with}
}
\value{
Table with derived model parameters
}
\description{
Return table with derived model parameters
}
