% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mslt}
\alias{print.mslt}
\title{Print mslt object}
\usage{
\method{print}{mslt}(x, ...)
}
\arguments{
\item{x}{Fitted model retuned by \code{\link{fitMSLT}}}

\item{...}{Extra argument}
}
\description{
Print mslt object
}
\details{
Print log-likelihood and the main convergence criteria
}
